CMakeLists.txt: spew lots of output
Spew lots of output, to make sure that the cmake-ng plugin can handle
it.  The cmake process is connected to the cmake-ng VM via a blocking
pipe, which can fill up if it's not being read.  (Similarly for "make",
but testing spewing a lot of make output is harder.)

Also get rid of a stray "WATERMELON."

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>